refactor: extract camera component
