+++ src/components/Camera/index.js
@@ -1,0 +1,1 @@
+export { default as Camera } from "./Camera";
